requirements.txt
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.26.0
pandas>=2.0.0
opencv-python>=4.8.0
mediapipe>=0.10.9
scikit-learn>=1.3.0
matplotlib>=3.8.0
seaborn>=0.13.0
gradio>=4.0.0
albumentations>=1.3.0
pillow>=10.0.0
tqdm>=4.66.0
joblib>=1.3.0
grad-cam==1.5.5
requests>=2.31.0